WIP: Player not bound to grid
This is a large set of changes required to unbind the player from the
grid. There are many more changes required, but this one gets many of
the pieces in the right spot. The game is not completely playable,
though, because player/item intersections are not yet working. This
means that most of the level ending mechanics are non-functional.

Also, the editor is not yet equipped to read and write levels in the new
format.

Much work to do, but this changes opens the door to smaller commits
henceforth.

+++ src/uitk/Animation.cpp
@@ -73,14 +73,14 @@ void Animation::draw(SDL_Renderer* renderer) {
     auto frameX  = this->frameList[frameXc];
     auto frameY  = this->frameList[frameYc];
 
-    SDL_Rect destination = {x, y, getWidth(), getHeight()};
+    SDL_Rect destination = {static_cast<uint16_t>(x), static_cast<uint16_t>(y), getWidth(), getHeight()};
     this->sheet->drawStill(renderer, frameX, frameY, destination);
 }
 
 /**
  * Move the drawing location of this Animation elsewhere.
  */
-void Animation::move(uint16_t x, uint16_t y) {
+void Animation::move(uint32_t x, uint32_t y) {
     this->x = x;
     this->y = y;
 }
