+++ src/uitk/Animation.hpp
@@ -10,7 +10,7 @@ public:
     Animation(SpriteSheet* spriteSheet, std::vector<uint16_t> frameList, uint16_t stillsPerSecond);
     bool advance();
     void draw(SDL_Renderer* renderer);
-    void move(uint16_t x, uint16_t y);
+    void move(uint32_t x, uint32_t y);
     bool isAnimating() const;
     void play();
     void stop();
@@ -22,8 +22,8 @@ private:
     uint16_t sps;
     std::vector<uint16_t> frameList;
     uint16_t currentStill;
-    uint16_t x;
-    uint16_t y;
+    uint32_t x;
+    uint32_t y;
     bool shouldAdvance;
 
     uint16_t framesPerStill;
